Add blueprint linting script
This should help us not ignore warnings in the blueprint syntax. Dredd
usually shows them, but will not fail the build.

We are using the `apiblueprint.org` API to do this, the linting will be skipped if
accessing the API somehow fails. Reasoning for that is not failing the
build when an unrelated service is down -- we'll catch the warnings
eventually.

The linting script fails the build when:
- there are unresolved references to partials, a sanity check because
  hercule doesn't set its exit status correctly.
- there are any warnings for the blueprint. In order to keep the script
  somewhat simple, we only display the message, not the location. Use
`drafter -l -u` locally for more specific information.
- there are any errors for the blueprint.